Reduce positive glucose effect if variable basal is used.
if keyForUseVariableBasal = true, then I'm using a too high basal rate. In that case also negative IOB should be reduced with the same percentage as used when reducing the basal if glucose value below average correction

this correction is not fully correct, it's just a small gain. It's to avoid that Loop starts overcorrecting with microbolusses, becuase it over estimates future glucose values, due to a too high basal rate.

The percentage is applied if current glucose value < 115

Also update README.md