Ongoing temp basal : IOB only for delivered units
When calculating IOB for temp basal that is still ongoing, and if it's a temp basal > scheduled basal, then calculate only the units that are (supposed to be) delivered, and not the units that will be delivered in the full duration.

This is only applicable to temp basal that are not yet finished (for finished temp basal, it was already the full amount obviously)

It's not 100% matching the real values, but at least it gives a better prediction.

This is to make sure that Loop predicts impact of units that are not yet delivered